requirements.txt
asgiref==3.5.0
autopep8==1.6.0
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-nested-routers==0.93.4
pycodestyle==2.8.0
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
